+++ ChangeLog.md
@@ -5,3 +5,4 @@
 ### Unreleased ###
 
 * Updates XD7WaitForSite to use common InvokeScriptBlock function
+* Added new XD7StoreFront resource